+++ GerasimovAPI/nameofproduct.cs
@@ -16,7 +16,7 @@ namespace GerasimovAPI
     {
         public int id { get; set; }
         public string name { get; set; }
-        public decimal price { get; set; }
+        public double price { get; set; }
         public string weight { get; set; }
         public string nameProiz { get; set; }
         public string countryProiz { get; set; }
